feat(elements): Expand debugging options [SDK-1472]
