+++ packages/elements/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@clerk/elements",
-  "version": "0.1.21",
+  "version": "0.1.22",
   "description": "Clerk Elements",
   "keywords": [
     "clerk",
@@ -51,9 +51,10 @@
     "dist"
   ],
   "scripts": {
-    "app:build": "(cd examples/nextjs && npm run build)",
-    "app:dev": "(cd examples/nextjs && npm run dev)",
-    "app:dev:debug": "(cd examples/nextjs && NEXT_PUBLIC_CLERK_ELEMENTS_DEBUG=true npm run dev)",
+    "app:build": "(cd examples/nextjs && npm run build --turbo)",
+    "app:dev": "(cd examples/nextjs && npm run dev --turbo)",
+    "app:dev:debug": "(cd examples/nextjs && NEXT_PUBLIC_CLERK_ELEMENTS_DEBUG=true npm run dev --turbo)",
+    "app:dev:debug:server": "(cd examples/nextjs && NEXT_PUBLIC_CLERK_ELEMENTS_DEBUG=true CLERK_ELEMENTS_DEBUG_SERVER=true npm run dev --turbo)",
     "app:e2e": "(cd examples/nextjs && npm run e2e)",
     "app:lint": "(cd examples/nextjs && npm run lint)",
     "build": "tsup",
