+++ packages/elements/src/react/hooks/index.ts
@@ -2,5 +2,4 @@ export { useActiveStates } from './use-active-states.hook';
 export { useActiveTags } from './use-active-tags.hook';
 export { useBrowserInspector } from './use-browser-inspector.hook';
 export { useThirdPartyProvider } from './use-third-party-provider.hook';
-export { consoleInspector, useConsoleInspector } from './use-console-inspector.hook';
 export { useFocus } from './use-focus.hook';
