--- packages/elements/src/react/hooks/use-console-inspector.hook.ts
@@ -1,7 +1,0 @@
-import { createConsoleInspector } from '~/react/utils/xstate';
-
-export const consoleInspector = createConsoleInspector();
-
-export function useConsoleInspector() {
-  return consoleInspector;
-}
